[Client] Handle sell trades properly for performance calculations
Sell trades now contribute to total holding performance via the profit
made from their sale. The 'cost' of the shares sold is calculated as a
weighted average of the previous buy trades.

Previous buy trades which have been 'sold' in a succeeding sell trade
have their performance figures removed so as not to double count.

There are various issues with this solution but it is good enough for now.